pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ad-mcquery"
version = "0.1.0"
description = "Ad analysis tool with image/video preprocessing and OCR"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Andy Khau"}
]
dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pytesseract>=0.3.10",
    "openai>=1.0.0",
    "ffmpeg-python==0.2.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.8.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false